moved valid_date to DateRange, all tests passing
